+++ ReLunacy.Engine/Scene/EntityManager.cs
@@ -2,7 +2,7 @@ using System.Numerics;
 using ReLunacy.Engine.Assets.Levels;
 using ReLunacy.Engine.Diagnostics;
 using ReLunacy.Engine.Rendering;
-using Veldrith;
+using NeoVeldrid;
 
 namespace ReLunacy.Engine.Scene;