+++ ReLunacy.Engine/Scene/EntityZone.cs
@@ -1,6 +1,6 @@
 using ReLunacy.Engine.Assets.Interfaces;
 using ReLunacy.Engine.Rendering;
-using Veldrith;
+using NeoVeldrid;
 
 namespace ReLunacy.Engine.Scene;